refactor: globals
