+++ crates/biome_js_analyze/src/globals/javascript/web.rs
@@ -1,38 +1,42 @@
 /// Sorted array of browser builtin
-pub const BROWSER: [&str; 825] = [
+///
+/// Source: https://github.com/sindresorhus/globals/blob/9e2e2598dabdb845ff76c0c3acf5c52c812a64de/globals.json#L368-L1442
+pub const BROWSER: &[&str; 1073] = &[
     "AbortController",
     "AbortSignal",
+    "AbsoluteOrientationSensor",
     "AbstractRange",
+    "Accelerometer",
     "AnalyserNode",
     "Animation",
     "AnimationEffect",
-    "AnimationEffectReadOnly",
-    "AnimationEffectTiming",
-    "AnimationEffectTimingReadOnly",
     "AnimationEvent",
     "AnimationPlaybackEvent",
     "AnimationTimeline",
-    "ApplicationCache",
-    "ApplicationCacheErrorEvent",
     "Attr",
     "Audio",
     "AudioBuffer",
     "AudioBufferSourceNode",
     "AudioContext",
+    "AudioData",
+    "AudioDecoder",
     "AudioDestinationNode",
+    "AudioEncoder",
     "AudioListener",
     "AudioNode",
     "AudioParam",
     "AudioParamMap",
     "AudioProcessingEvent",
     "AudioScheduledSourceNode",
+    "AudioSinkInfo",
     "AudioWorklet",
-    "AudioWorkletGlobalScope",
     "AudioWorkletNode",
-    "AudioWorkletProcessor",
     "AuthenticatorAssertionResponse",
     "AuthenticatorAttestationResponse",
     "AuthenticatorResponse",
+    "BackgroundFetchManager",
+    "BackgroundFetchRecord",
+    "BackgroundFetchRegistration",
     "BarProp",
     "BaseAudioContext",
     "BatteryManager",
@@ -40,8 +44,16 @@ pub const BROWSER: [&str; 825] = [
     "BiquadFilterNode",
     "Blob",
     "BlobEvent",
+    "Bluetooth",
+    "BluetoothCharacteristicProperties",
+    "BluetoothDevice",
+    "BluetoothRemoteGATTCharacteristic",
+    "BluetoothRemoteGATTDescriptor",
+    "BluetoothRemoteGATTServer",
+    "BluetoothRemoteGATTService",
+    "BluetoothUUID",
     "BroadcastChannel",
-    "BudgetService",
+    "BrowserCaptureMediaStreamTrack",
     "ByteLengthQueuingStrategy",
     "CDATASection",
     "CSS",
@@ -50,6 +62,7 @@ pub const BROWSER: [&str; 825] = [
     "CSSContainerRule",
     "CSSCounterStyleRule",
     "CSSFontFaceRule",
+    "CSSFontFeatureValuesRule",
     "CSSFontPaletteValuesRule",
     "CSSGroupingRule",
     "CSSImageValue",
@@ -74,14 +87,17 @@ pub const BROWSER: [&str; 825] = [
     "CSSNumericValue",
     "CSSPageRule",
     "CSSPerspective",
+    "CSSPositionValue",
     "CSSPropertyRule",
     "CSSRotate",
     "CSSRule",
     "CSSRuleList",
     "CSSScale",
+    "CSSScopeRule",
     "CSSSkew",
     "CSSSkewX",
     "CSSSkewY",
+    "CSSStartingStyleRule",
     "CSSStyleDeclaration",
     "CSSStyleRule",
     "CSSStyleSheet",
@@ -96,12 +112,16 @@ pub const BROWSER: [&str; 825] = [
     "CSSVariableReferenceValue",
     "Cache",
     "CacheStorage",
+    "CanvasCaptureMediaStream",
     "CanvasCaptureMediaStreamTrack",
     "CanvasGradient",
     "CanvasPattern",
     "CanvasRenderingContext2D",
+    "CaptureController",
+    "CaretPosition",
     "ChannelMergerNode",
     "ChannelSplitterNode",
+    "CharacterBoundsUpdateEvent",
     "CharacterData",
     "Clipboard",
     "ClipboardEvent",
@@ -111,14 +131,20 @@ pub const BROWSER: [&str; 825] = [
     "CompositionEvent",
     "CompressionStream",
     "ConstantSourceNode",
+    "ContentVisibilityAutoStateChangeEvent",
     "ConvolverNode",
+    "CookieChangeEvent",
+    "CookieStore",
+    "CookieStoreManager",
     "CountQueuingStrategy",
     "Credential",
     "CredentialsContainer",
+    "CropTarget",
     "Crypto",
     "CryptoKey",
     "CustomElementRegistry",
     "CustomEvent",
+    "CustomStateSet",
     "DOMError",
     "DOMException",
     "DOMImplementation",
@@ -139,18 +165,23 @@ pub const BROWSER: [&str; 825] = [
     "DataTransferItemList",
     "DecompressionStream",
     "DelayNode",
+    "DelegatedInkTrailPresenter",
     "DeviceMotionEvent",
     "DeviceMotionEventAcceleration",
     "DeviceMotionEventRotationRate",
     "DeviceOrientationEvent",
     "Document",
     "DocumentFragment",
+    "DocumentPictureInPicture",
+    "DocumentPictureInPictureEvent",
     "DocumentTimeline",
     "DocumentType",
     "DragEvent",
     "DynamicsCompressorNode",
+    "EditContext",
     "Element",
     "ElementInternals",
+    "EncodedAudioChunk",
     "EncodedVideoChunk",
     "ErrorEvent",
     "Event",
@@ -158,27 +189,87 @@ pub const BROWSER: [&str; 825] = [
     "EventSource",
     "EventTarget",
     "External",
+    "EyeDropper",
+    "FeaturePolicy",
+    "FederatedCredential",
     "File",
     "FileList",
     "FileReader",
+    "FileSystem",
+    "FileSystemDirectoryEntry",
     "FileSystemDirectoryHandle",
+    "FileSystemDirectoryReader",
+    "FileSystemEntry",
+    "FileSystemFileEntry",
     "FileSystemFileHandle",
     "FileSystemHandle",
     "FileSystemWritableFileStream",
     "FocusEvent",
+    "FontData",
     "FontFace",
+    "FontFaceSet",
     "FontFaceSetLoadEvent",
     "FormData",
     "FormDataEvent",
+    "FragmentDirective",
+    "GPU",
+    "GPUAdapter",
+    "GPUAdapterInfo",
+    "GPUBindGroup",
+    "GPUBindGroupLayout",
+    "GPUBuffer",
+    "GPUBufferUsage",
+    "GPUCanvasContext",
+    "GPUColorWrite",
+    "GPUCommandBuffer",
+    "GPUCommandEncoder",
+    "GPUCompilationInfo",
+    "GPUCompilationMessage",
+    "GPUComputePassEncoder",
+    "GPUComputePipeline",
+    "GPUDevice",
+    "GPUDeviceLostInfo",
+    "GPUError",
+    "GPUExternalTexture",
+    "GPUInternalError",
+    "GPUMapMode",
+    "GPUOutOfMemoryError",
+    "GPUPipelineError",
+    "GPUPipelineLayout",
+    "GPUQuerySet",
+    "GPUQueue",
+    "GPURenderBundle",
+    "GPURenderBundleEncoder",
+    "GPURenderPassEncoder",
+    "GPURenderPipeline",
+    "GPUSampler",
+    "GPUShaderModule",
+    "GPUShaderStage",
+    "GPUSupportedFeatures",
+    "GPUSupportedLimits",
+    "GPUTexture",
+    "GPUTextureUsage",
+    "GPUTextureView",
+    "GPUUncapturedErrorEvent",
+    "GPUValidationError",
     "GainNode",
     "Gamepad",
+    "GamepadAxisMoveEvent",
     "GamepadButton",
+    "GamepadButtonEvent",
     "GamepadEvent",
     "GamepadHapticActuator",
+    "GamepadPose",
     "Geolocation",
     "GeolocationCoordinates",
     "GeolocationPosition",
     "GeolocationPositionError",
+    "GravitySensor",
+    "Gyroscope",
+    "HID",
+    "HIDConnectionEvent",
+    "HIDDevice",
+    "HIDInputReportEvent",
     "HTMLAllCollection",
     "HTMLAnchorElement",
     "HTMLAreaElement",
@@ -189,7 +280,6 @@ pub const BROWSER: [&str; 825] = [
     "HTMLButtonElement",
     "HTMLCanvasElement",
     "HTMLCollection",
-    "HTMLContentElement",
     "HTMLDListElement",
     "HTMLDataElement",
     "HTMLDataListElement",
@@ -238,7 +328,6 @@ pub const BROWSER: [&str; 825] = [
     "HTMLQuoteElement",
     "HTMLScriptElement",
     "HTMLSelectElement",
-    "HTMLShadowElement",
     "HTMLSlotElement",
     "HTMLSourceElement",
     "HTMLSpanElement",
@@ -259,6 +348,8 @@ pub const BROWSER: [&str; 825] = [
     "HTMLVideoElement",
     "HashChangeEvent",
     "Headers",
+    "Highlight",
+    "HighlightRegistry",
     "History",
     "IDBCursor",
     "IDBCursorWithValue",
@@ -272,20 +363,36 @@ pub const BROWSER: [&str; 825] = [
     "IDBTransaction",
     "IDBVersionChangeEvent",
     "IIRFilterNode",
+    "IdentityCredential",
+    "IdentityCredentialError",
+    "IdentityProvider",
     "IdleDeadline",
+    "IdleDetector",
     "Image",
     "ImageBitmap",
     "ImageBitmapRenderingContext",
     "ImageCapture",
     "ImageData",
+    "ImageDecoder",
+    "ImageTrack",
+    "ImageTrackList",
+    "Ink",
+    "InputDeviceCapabilities",
     "InputDeviceInfo",
     "InputEvent",
     "IntersectionObserver",
     "IntersectionObserverEntry",
-    "Intl",
+    "Iterator",
+    "Keyboard",
     "KeyboardEvent",
+    "KeyboardLayoutMap",
     "KeyframeEffect",
-    "KeyframeEffectReadOnly",
+    "LargestContentfulPaint",
+    "LaunchParams",
+    "LaunchQueue",
+    "LayoutShift",
+    "LayoutShiftAttribution",
+    "LinearAccelerationSensor",
     "Location",
     "Lock",
     "LockManager",
@@ -299,11 +406,13 @@ pub const BROWSER: [&str; 825] = [
     "MIDIPort",
     "MathMLElement",
     "MediaCapabilities",
+    "MediaCapabilitiesInfo",
     "MediaDeviceInfo",
     "MediaDevices",
     "MediaElementAudioSourceNode",
     "MediaEncryptedEvent",
     "MediaError",
+    "MediaKeyError",
     "MediaKeyMessageEvent",
     "MediaKeySession",
     "MediaKeyStatusMap",
@@ -314,15 +423,20 @@ pub const BROWSER: [&str; 825] = [
     "MediaQueryList",
     "MediaQueryListEvent",
     "MediaRecorder",
+    "MediaRecorderErrorEvent",
     "MediaSession",
-    "MediaSettingsRange",
     "MediaSource",
+    "MediaSourceHandle",
     "MediaStream",
     "MediaStreamAudioDestinationNode",
     "MediaStreamAudioSourceNode",
     "MediaStreamEvent",
     "MediaStreamTrack",
+    "MediaStreamTrackAudioSourceNode",
     "MediaStreamTrackEvent",
+    "MediaStreamTrackGenerator",
+    "MediaStreamTrackProcessor",
+    "MediaStreamTrackVideoStats",
     "MessageChannel",
     "MessageEvent",
     "MessagePort",
@@ -333,30 +447,46 @@ pub const BROWSER: [&str; 825] = [
     "MutationObserver",
     "MutationRecord",
     "NamedNodeMap",
+    "NavigateEvent",
+    "Navigation",
+    "NavigationCurrentEntryChangeEvent",
+    "NavigationDestination",
+    "NavigationHistoryEntry",
     "NavigationPreloadManager",
+    "NavigationTransition",
     "Navigator",
+    "NavigatorLogin",
+    "NavigatorManagedData",
+    "NavigatorUAData",
     "NetworkInformation",
     "Node",
     "NodeFilter",
     "NodeIterator",
     "NodeList",
     "Notification",
+    "NotifyPaintEvent",
+    "OTPCredential",
     "OfflineAudioCompletionEvent",
     "OfflineAudioContext",
     "OffscreenCanvas",
     "OffscreenCanvasRenderingContext2D",
     "Option",
+    "OrientationSensor",
     "OscillatorNode",
     "OverconstrainedError",
+    "PERSISTENT",
     "PageTransitionEvent",
     "PannerNode",
+    "PasswordCredential",
     "Path2D",
     "PaymentAddress",
+    "PaymentManager",
     "PaymentMethodChangeEvent",
     "PaymentRequest",
     "PaymentRequestUpdateEvent",
     "PaymentResponse",
     "Performance",
+    "PerformanceElementTiming",
     "PerformanceEntry",
     "PerformanceEventTiming",
     "PerformanceLongTaskTiming",
@@ -370,10 +500,10 @@ pub const BROWSER: [&str; 825] = [
     "PerformanceResourceTiming",
     "PerformanceServerTiming",
     "PerformanceTiming",
+    "PeriodicSyncManager",
     "PeriodicWave",
     "PermissionStatus",
     "Permissions",
-    "PhotoCapabilities",
     "PictureInPictureEvent",
     "PictureInPictureWindow",
     "Plugin",
@@ -389,6 +519,7 @@ pub const BROWSER: [&str; 825] = [
     "PresentationReceiver",
     "PresentationRequest",
     "ProcessingInstruction",
+    "Profiler",
     "ProgressEvent",
     "PromiseRejectionEvent",
     "PublicKeyCredential",
@@ -406,13 +537,12 @@ pub const BROWSER: [&str; 825] = [
     "RTCError",
     "RTCErrorEvent",
     "RTCIceCandidate",
-    "RTCIceGatherer",
     "RTCIceTransport",
     "RTCPeerConnection",
     "RTCPeerConnectionIceErrorEvent",
     "RTCPeerConnectionIceEvent",
-    "RTCRtpContributingSource",
     "RTCRtpReceiver",
+    "RTCRtpScriptTransform",
     "RTCRtpSender",
     "RTCRtpTransceiver",
     "RTCSctpTransport",
@@ -427,6 +557,7 @@ pub const BROWSER: [&str; 825] = [
     "ReadableStreamBYOBRequest",
     "ReadableStreamDefaultController",
     "ReadableStreamDefaultReader",
+    "RelativeOrientationSensor",
     "RemotePlayback",
     "ReportingObserver",
     "Request",
@@ -457,7 +588,6 @@ pub const BROWSER: [&str; 825] = [
     "SVGComponentTransferFunctionElement",
     "SVGDefsElement",
     "SVGDescElement",
-    "SVGDiscardElement",
     "SVGElement",
     "SVGEllipseElement",
     "SVGFEBlendElement",
@@ -532,11 +662,20 @@ pub const BROWSER: [&str; 825] = [
     "SVGUnitTypes",
     "SVGUseElement",
     "SVGViewElement",
+    "Scheduler",
+    "Scheduling",
     "Screen",
+    "ScreenDetailed",
+    "ScreenDetails",
     "ScreenOrientation",
     "ScriptProcessorNode",
+    "ScrollTimeline",
     "SecurityPolicyViolationEvent",
     "Selection",
+    "Sensor",
+    "SensorErrorEvent",
+    "Serial",
+    "SerialPort",
     "ServiceWorker",
     "ServiceWorkerContainer",
     "ServiceWorkerRegistration",
@@ -544,12 +683,16 @@ pub const BROWSER: [&str; 825] = [
     "SharedWorker",
     "SourceBuffer",
     "SourceBufferList",
+    "SpeechSynthesis",
     "SpeechSynthesisErrorEvent",
     "SpeechSynthesisEvent",
     "SpeechSynthesisUtterance",
+    "SpeechSynthesisVoice",
     "StaticRange",
     "StereoPannerNode",
     "Storage",
+    "StorageBucket",
+    "StorageBucketManager",
     "StorageEvent",
     "StorageManager",
     "StylePropertyMap",
@@ -558,18 +701,27 @@ pub const BROWSER: [&str; 825] = [
     "StyleSheetList",
     "SubmitEvent",
     "SubtleCrypto",
+    "SyncManager",
+    "TEMPORARY",
     "TaskAttributionTiming",
+    "TaskController",
+    "TaskPriorityChangeEvent",
+    "TaskSignal",
     "Text",
     "TextDecoder",
     "TextDecoderStream",
     "TextEncoder",
     "TextEncoderStream",
     "TextEvent",
+    "TextFormat",
+    "TextFormatUpdateEvent",
     "TextMetrics",
     "TextTrack",
     "TextTrackCue",
     "TextTrackCueList",
     "TextTrackList",
+    "TextUpdateEvent",
+    "TimeEvent",
     "TimeRanges",
     "ToggleEvent",
     "Touch",
@@ -580,18 +732,44 @@ pub const BROWSER: [&str; 825] = [
     "TransformStreamDefaultController",
     "TransitionEvent",
     "TreeWalker",
+    "TrustedHTML",
+    "TrustedScript",
+    "TrustedScriptURL",
+    "TrustedTypePolicy",
+    "TrustedTypePolicyFactory",
     "UIEvent",
     "URL",
+    "URLPattern",
     "URLSearchParams",
+    "USB",
+    "USBAlternateInterface",
+    "USBConfiguration",
+    "USBConnectionEvent",
+    "USBDevice",
+    "USBEndpoint",
+    "USBInTransferResult",
+    "USBInterface",
+    "USBIsochronousInTransferPacket",
+    "USBIsochronousInTransferResult",
+    "USBIsochronousOutTransferPacket",
+    "USBIsochronousOutTransferResult",
+    "USBOutTransferResult",
     "UserActivation",
     "VTTCue",
+    "VTTRegion",
     "ValidityState",
     "VideoColorSpace",
     "VideoDecoder",
     "VideoEncoder",
     "VideoFrame",
     "VideoPlaybackQuality",
+    "ViewTimeline",
+    "ViewTransition",
+    "VirtualKeyboard",
+    "VirtualKeyboardGeometryChangeEvent",
+    "VisibilityStateEntry",
     "VisualViewport",
+    "WGSLLanguageFeatures",
     "WakeLock",
     "WakeLockSentinel",
     "WaveShaperNode",
@@ -613,14 +791,17 @@ pub const BROWSER: [&str; 825] = [
     "WebGLTransformFeedback",
     "WebGLUniformLocation",
     "WebGLVertexArrayObject",
-    "WebKitCSSMatrix",
     "WebSocket",
     "WebTransport",
     "WebTransportBidirectionalStream",
     "WebTransportDatagramDuplexStream",
     "WebTransportError",
+    "WebTransportReceiveStream",
+    "WebTransportSendStream",
     "WheelEvent",
     "Window",
+    "WindowControlsOverlay",
+    "WindowControlsOverlayGeometryChangeEvent",
     "Worker",
     "Worklet",
     "WritableStream",
@@ -634,10 +815,44 @@ pub const BROWSER: [&str; 825] = [
     "XPathEvaluator",
     "XPathExpression",
     "XPathResult",
+    "XRAnchor",
+    "XRAnchorSet",
+    "XRBoundedReferenceSpace",
+    "XRCPUDepthInformation",
+    "XRCamera",
+    "XRDOMOverlayState",
+    "XRDepthInformation",
+    "XRFrame",
+    "XRHitTestResult",
+    "XRHitTestSource",
+    "XRInputSource",
+    "XRInputSourceArray",
+    "XRInputSourceEvent",
+    "XRInputSourcesChangeEvent",
+    "XRLayer",
+    "XRLightEstimate",
+    "XRLightProbe",
+    "XRPose",
+    "XRRay",
+    "XRReferenceSpace",
+    "XRReferenceSpaceEvent",
+    "XRRenderState",
+    "XRRigidTransform",
+    "XRSession",
+    "XRSessionEvent",
+    "XRSpace",
+    "XRSystem",
+    "XRTransientInputHitTestResult",
+    "XRTransientInputHitTestSource",
+    "XRView",
+    "XRViewerPose",
+    "XRViewport",
+    "XRWebGLBinding",
+    "XRWebGLDepthInformation",
+    "XRWebGLLayer",
     "XSLTProcessor",
     "addEventListener",
     "alert",
-    "applicationCache",
     "atob",
     "blur",
     "btoa",
@@ -651,14 +866,16 @@ pub const BROWSER: [&str; 825] = [
     "closed",
     "confirm",
     "console",
+    "cookieStore",
     "createImageBitmap",
+    "credentialless",
+    "crossOriginIsolated",
     "crypto",
     "customElements",
-    "defaultStatus",
-    "defaultstatus",
     "devicePixelRatio",
     "dispatchEvent",
     "document",
+    "documentPictureInPicture",
     "event",
     "external",
     "fetch",
@@ -667,12 +884,14 @@ pub const BROWSER: [&str; 825] = [
     "frameElement",
     "frames",
     "getComputedStyle",
+    "getScreenDetails",
     "getSelection",
     "history",
     "indexedDB",
     "innerHeight",
     "innerWidth",
     "isSecureContext",
+    "launchQueue",
     "length",
     "localStorage",
     "location",
@@ -682,18 +901,24 @@ pub const BROWSER: [&str; 825] = [
     "moveBy",
     "moveTo",
     "name",
+    "navigation",
     "navigator",
     "offscreenBuffering",
     "onabort",
     "onafterprint",
+    "onanimationcancel",
     "onanimationend",
     "onanimationiteration",
     "onanimationstart",
     "onappinstalled",
     "onauxclick",
+    "onbeforeinput",
     "onbeforeinstallprompt",
+    "onbeforematch",
     "onbeforeprint",
+    "onbeforetoggle",
     "onbeforeunload",
+    "onbeforexrselect",
     "onblur",
     "oncancel",
     "oncanplay",
@@ -701,8 +926,13 @@ pub const BROWSER: [&str; 825] = [
     "onchange",
     "onclick",
     "onclose",
+    "oncontentvisibilityautostatechange",
+    "oncontextlost",
     "oncontextmenu",
+    "oncontextrestored",
+    "oncopy",
     "oncuechange",
+    "oncut",
     "ondblclick",
     "ondevicemotion",
     "ondeviceorientation",
@@ -719,6 +949,9 @@ pub const BROWSER: [&str; 825] = [
     "onended",
     "onerror",
     "onfocus",
+    "onformdata",
+    "ongamepadconnected",
+    "ongamepaddisconnected",
     "ongotpointercapture",
     "onhashchange",
     "oninput",
@@ -746,6 +979,7 @@ pub const BROWSER: [&str; 825] = [
     "ononline",
     "onpagehide",
     "onpageshow",
+    "onpaste",
     "onpause",
     "onplay",
     "onplaying",
@@ -756,6 +990,7 @@ pub const BROWSER: [&str; 825] = [
     "onpointermove",
     "onpointerout",
     "onpointerover",
+    "onpointerrawupdate",
     "onpointerup",
     "onpopstate",
     "onprogress",
@@ -764,26 +999,34 @@ pub const BROWSER: [&str; 825] = [
     "onreset",
     "onresize",
     "onscroll",
+    "onscrollend",
     "onsearch",
+    "onsecuritypolicyviolation",
     "onseeked",
     "onseeking",
     "onselect",
+    "onselectionchange",
+    "onselectstart",
+    "onslotchange",
     "onstalled",
     "onstorage",
     "onsubmit",
     "onsuspend",
     "ontimeupdate",
     "ontoggle",
+    "ontransitioncancel",
     "ontransitionend",
+    "ontransitionrun",
+    "ontransitionstart",
     "onunhandledrejection",
     "onunload",
     "onvolumechange",
     "onwaiting",
     "onwheel",
     "open",
-    "openDatabase",
     "opener",
     "origin",
+    "originAgentCluster",
     "outerHeight",
     "outerWidth",
     "pageXOffset",
@@ -794,13 +1037,15 @@ pub const BROWSER: [&str; 825] = [
     "postMessage",
     "print",
     "prompt",
+    "queryLocalFonts",
     "queueMicrotask",
-    "registerProcessor",
     "removeEventListener",
+    "reportError",
     "requestAnimationFrame",
     "requestIdleCallback",
     "resizeBy",
     "resizeTo",
+    "scheduler",
     "screen",
     "screenLeft",
     "screenTop",
@@ -816,27 +1061,45 @@ pub const BROWSER: [&str; 825] = [
     "sessionStorage",
     "setInterval",
     "setTimeout",
+    "showDirectoryPicker",
+    "showOpenFilePicker",
+    "showSaveFilePicker",
     "speechSynthesis",
     "status",
     "statusbar",
     "stop",
+    "structuredClone",
     "styleMedia",
     "toolbar",
     "top",
+    "trustedTypes",
     "visualViewport",
     "window",
 ];
 
-pub const SERVICE_WORKER: [&str; 93] = [
+/// Sorted array of browser service worker builtin
+///
+/// Source: https://github.com/sindresorhus/globals/blob/9e2e2598dabdb845ff76c0c3acf5c52c812a64de/globals.json#L2212-L2334
+pub const SERVICE_WORKER: &[&str; 121] = &[
     "Blob",
     "BroadcastChannel",
+    "ByteLengthQueuingStrategy",
     "Cache",
     "CacheStorage",
     "Client",
     "Clients",
+    "CompressionStream",
+    "CountQueuingStrategy",
+    "Crypto",
+    "CryptoKey",
+    "CustomEvent",
+    "DecompressionStream",
+    "ErrorEvent",
+    "Event",
     "ExtendableEvent",
     "ExtendableMessageEvent",
     "FetchEvent",
+    "File",
     "FileReaderSync",
     "FormData",
     "Headers",
@@ -853,6 +1116,7 @@ pub const SERVICE_WORKER: [&str; 93] = [
     "IDBVersionChangeEvent",
     "ImageData",
     "MessageChannel",
+    "MessageEvent",
     "MessagePort",
     "Notification",
     "Performance",
@@ -860,9 +1124,16 @@ pub const SERVICE_WORKER: [&str; 93] = [
     "PerformanceMark",
     "PerformanceMeasure",
     "PerformanceNavigation",
+    "PerformanceObserver",
+    "PerformanceObserverEntryList",
     "PerformanceResourceTiming",
     "PerformanceTiming",
-    "Promise",
+    "ReadableByteStreamController",
+    "ReadableStream",
+    "ReadableStreamBYOBReader",
+    "ReadableStreamBYOBRequest",
+    "ReadableStreamDefaultController",
+    "ReadableStreamDefaultReader",
     "Request",
     "Response",
     "ServiceWorker",
@@ -870,14 +1141,23 @@ pub const SERVICE_WORKER: [&str; 93] = [
     "ServiceWorkerGlobalScope",
     "ServiceWorkerMessageEvent",
     "ServiceWorkerRegistration",
+    "SubtleCrypto",
     "TextDecoder",
+    "TextDecoderStream",
     "TextEncoder",
+    "TextEncoderStream",
+    "TransformStream",
+    "TransformStreamDefaultController",
     "URL",
     "URLSearchParams",
+    "WebAssembly",
     "WebSocket",
     "WindowClient",
     "Worker",
     "WorkerGlobalScope",
+    "WritableStream",
+    "WritableStreamDefaultController",
+    "WritableStreamDefaultWriter",
     "XMLHttpRequest",
     "addEventListener",
     "applicationCache",
@@ -889,6 +1169,7 @@ pub const SERVICE_WORKER: [&str; 93] = [
     "clients",
     "close",
     "console",
+    "crypto",
     "fetch",
     "importScripts",
     "indexedDB",
@@ -923,12 +1204,102 @@ pub const SERVICE_WORKER: [&str; 93] = [
     "skipWaiting",
 ];
 
-pub const WORKER: [&str; 71] = [
+/// Sorted array of browser web worker builtin
+///
+/// Source: https://github.com/sindresorhus/globals/blob/9e2e2598dabdb845ff76c0c3acf5c52c812a64de/globals.json#L1443-L1726
+pub const WEB_WORKER: &[&str; 282] = &[
+    "AbortController",
+    "AbortSignal",
+    "AudioData",
+    "AudioDecoder",
+    "AudioEncoder",
+    "BackgroundFetchManager",
+    "BackgroundFetchRecord",
+    "BackgroundFetchRegistration",
     "Blob",
     "BroadcastChannel",
+    "ByteLengthQueuingStrategy",
+    "CSSSkewX",
+    "CSSSkewY",
     "Cache",
+    "CacheStorage",
+    "CanvasGradient",
+    "CanvasPattern",
+    "CloseEvent",
+    "CompressionStream",
+    "CountQueuingStrategy",
+    "CropTarget",
+    "Crypto",
+    "CryptoKey",
+    "CustomEvent",
+    "DOMException",
+    "DOMMatrix",
+    "DOMMatrixReadOnly",
+    "DOMPoint",
+    "DOMPointReadOnly",
+    "DOMQuad",
+    "DOMRect",
+    "DOMRectReadOnly",
+    "DOMStringList",
+    "DecompressionStream",
+    "DedicatedWorkerGlobalScope",
+    "EncodedAudioChunk",
+    "EncodedVideoChunk",
+    "ErrorEvent",
+    "Event",
+    "EventSource",
+    "EventTarget",
+    "File",
+    "FileList",
+    "FileReader",
     "FileReaderSync",
+    "FileSystemDirectoryHandle",
+    "FileSystemFileHandle",
+    "FileSystemHandle",
+    "FileSystemSyncAccessHandle",
+    "FileSystemWritableFileStream",
+    "FontFace",
     "FormData",
+    "GPU",
+    "GPUAdapter",
+    "GPUAdapterInfo",
+    "GPUBindGroup",
+    "GPUBindGroupLayout",
+    "GPUBuffer",
+    "GPUBufferUsage",
+    "GPUCanvasContext",
+    "GPUColorWrite",
+    "GPUCommandBuffer",
+    "GPUCommandEncoder",
+    "GPUCompilationInfo",
+    "GPUCompilationMessage",
+    "GPUComputePassEncoder",
+    "GPUComputePipeline",
+    "GPUDevice",
+    "GPUDeviceLostInfo",
+    "GPUError",
+    "GPUExternalTexture",
+    "GPUInternalError",
+    "GPUMapMode",
+    "GPUOutOfMemoryError",
+    "GPUPipelineError",
+    "GPUPipelineLayout",
+    "GPUQuerySet",
+    "GPUQueue",
+    "GPURenderBundle",
+    "GPURenderBundleEncoder",
+    "GPURenderPassEncoder",
+    "GPURenderPipeline",
+    "GPUSampler",
+    "GPUShaderModule",
+    "GPUShaderStage",
+    "GPUSupportedFeatures",
+    "GPUSupportedLimits",
+    "GPUTexture",
+    "GPUTextureUsage",
+    "GPUTextureView",
+    "GPUUncapturedErrorEvent",
+    "GPUValidationError",
     "Headers",
     "IDBCursor",
     "IDBCursorWithValue",
@@ -941,65 +1312,208 @@ pub const WORKER: [&str; 71] = [
     "IDBRequest",
     "IDBTransaction",
     "IDBVersionChangeEvent",
+    "IdleDetector",
+    "ImageBitmap",
+    "ImageBitmapRenderingContext",
     "ImageData",
+    "ImageDecoder",
+    "ImageTrack",
+    "ImageTrackList",
+    "Iterator",
+    "Lock",
+    "LockManager",
+    "MediaCapabilities",
+    "MediaSource",
+    "MediaSourceHandle",
     "MessageChannel",
+    "MessageEvent",
     "MessagePort",
+    "NavigationPreloadManager",
+    "NavigatorUAData",
+    "NetworkInformation",
     "Notification",
+    "OffscreenCanvas",
+    "OffscreenCanvasRenderingContext2D",
+    "PERSISTENT",
+    "Path2D",
     "Performance",
     "PerformanceEntry",
     "PerformanceMark",
     "PerformanceMeasure",
-    "PerformanceNavigation",
+    "PerformanceObserver",
+    "PerformanceObserverEntryList",
     "PerformanceResourceTiming",
-    "PerformanceTiming",
-    "Promise",
+    "PerformanceServerTiming",
+    "PeriodicSyncManager",
+    "PermissionStatus",
+    "Permissions",
+    "ProgressEvent",
+    "PromiseRejectionEvent",
+    "PushManager",
+    "PushSubscription",
+    "PushSubscriptionOptions",
+    "RTCEncodedAudioFrame",
+    "RTCEncodedVideoFrame",
+    "ReadableByteStreamController",
+    "ReadableStream",
+    "ReadableStreamBYOBReader",
+    "ReadableStreamBYOBRequest",
+    "ReadableStreamDefaultController",
+    "ReadableStreamDefaultReader",
+    "ReportingObserver",
     "Request",
     "Response",
+    "Scheduler",
+    "SecurityPolicyViolationEvent",
+    "Serial",
+    "SerialPort",
     "ServiceWorkerRegistration",
+    "SourceBuffer",
+    "SourceBufferList",
+    "StorageBucket",
+    "StorageBucketManager",
+    "StorageManager",
+    "SubtleCrypto",
+    "SyncManager",
+    "TEMPORARY",
+    "TaskController",
+    "TaskPriorityChangeEvent",
+    "TaskSignal",
     "TextDecoder",
+    "TextDecoderStream",
     "TextEncoder",
+    "TextEncoderStream",
+    "TextMetrics",
+    "TransformStream",
+    "TransformStreamDefaultController",
+    "TrustedHTML",
+    "TrustedScript",
+    "TrustedScriptURL",
+    "TrustedTypePolicy",
+    "TrustedTypePolicyFactory",
     "URL",
+    "URLPattern",
     "URLSearchParams",
+    "USB",
+    "USBAlternateInterface",
+    "USBConfiguration",
+    "USBConnectionEvent",
+    "USBDevice",
+    "USBEndpoint",
+    "USBInTransferResult",
+    "USBInterface",
+    "USBIsochronousInTransferPacket",
+    "USBIsochronousInTransferResult",
+    "USBIsochronousOutTransferPacket",
+    "USBIsochronousOutTransferResult",
+    "USBOutTransferResult",
+    "UserActivation",
+    "VideoColorSpace",
+    "VideoDecoder",
+    "VideoEncoder",
+    "VideoFrame",
+    "WGSLLanguageFeatures",
+    "WebAssembly",
+    "WebGL2RenderingContext",
+    "WebGLActiveInfo",
+    "WebGLBuffer",
+    "WebGLContextEvent",
+    "WebGLFramebuffer",
+    "WebGLProgram",
+    "WebGLQuery",
+    "WebGLRenderbuffer",
+    "WebGLRenderingContext",
+    "WebGLSampler",
+    "WebGLShader",
+    "WebGLShaderPrecisionFormat",
+    "WebGLSync",
+    "WebGLTexture",
+    "WebGLTransformFeedback",
+    "WebGLUniformLocation",
+    "WebGLVertexArrayObject",
     "WebSocket",
+    "WebTransport",
+    "WebTransportBidirectionalStream",
+    "WebTransportDatagramDuplexStream",
+    "WebTransportError",
     "Worker",
     "WorkerGlobalScope",
+    "WorkerLocation",
+    "WorkerNavigator",
+    "WritableStream",
+    "WritableStreamDefaultController",
+    "WritableStreamDefaultWriter",
     "XMLHttpRequest",
+    "XMLHttpRequestEventTarget",
+    "XMLHttpRequestUpload",
     "addEventListener",
-    "applicationCache",
     "atob",
     "btoa",
     "caches",
+    "cancelAnimationFrame",
     "clearInterval",
     "clearTimeout",
     "close",
     "console",
+    "createImageBitmap",
+    "crossOriginIsolated",
+    "crypto",
+    "dispatchEvent",
     "fetch",
+    "fonts",
     "importScripts",
     "indexedDB",
+    "isSecureContext",
     "location",
     "name",
     "navigator",
-    "onclose",
-    "onconnect",
     "onerror",
     "onlanguagechange",
     "onmessage",
-    "onoffline",
-    "ononline",
+    "onmessageerror",
     "onrejectionhandled",
     "onunhandledrejection",
+    "origin",
     "performance",
     "postMessage",
     "queueMicrotask",
     "removeEventListener",
+    "reportError",
+    "requestAnimationFrame",
+    "scheduler",
     "self",
     "setInterval",
     "setTimeout",
+    "structuredClone",
+    "trustedTypes",
+    "webkitRequestFileSystem",
+    "webkitRequestFileSystemSync",
+    "webkitResolveLocalFileSystemSyncURL",
+    "webkitResolveLocalFileSystemURL",
 ];
+
+/// Returns `true` if `name` is a web global
+///
+/// ```
+/// use biome_js_analyze::globals::javascript::web::is_global;
+///
+/// assert!(is_global(&"AbortController"));
+/// ```
+pub fn is_global(name: &str) -> bool {
+    BROWSER.binary_search(&name).is_ok()
+        || WEB_WORKER.binary_search(&name).is_ok()
+        || SERVICE_WORKER.binary_search(&name).is_ok()
+}
 
 #[test]
 fn test_order() {
     for items in BROWSER.windows(2) {
+        assert!(items[0] < items[1], "{} < {}", items[0], items[1]);
+    }
+    for items in SERVICE_WORKER.windows(2) {
+        assert!(items[0] < items[1], "{} < {}", items[0], items[1]);
+    }
+    for items in WEB_WORKER.windows(2) {
         assert!(items[0] < items[1], "{} < {}", items[0], items[1]);
     }
 }
