+++ crates/biome_js_analyze/src/globals/module/mod.rs
@@ -1,0 +1,3 @@
+pub mod node;
+
+pub use node::is_builtin_module as is_node_builtin_module;
