--- crates/biome_js_analyze/src/globals/typescript.rs
@@ -1,975 +1,0 @@
-/// Sorted array of TypeScript builtin
-pub const TYPESCRIPT_BUILTIN: [&str; 965] = [
-    "ANGLE_instanced_arrays",
-    "ARIAMixin",
-    "AbortSignalEventMap",
-    "AbstractWorker",
-    "AbstractWorkerEventMap",
-    "AddEventListenerOptions",
-    "AesCbcParams",
-    "AesCtrParams",
-    "AesDerivedKeyParams",
-    "AesGcmParams",
-    "AesKeyAlgorithm",
-    "AesKeyGenParams",
-    "AggregateErrorConstructor",
-    "Algorithm",
-    "AlgorithmIdentifier",
-    "AlignSetting",
-    "AllowSharedBufferSource",
-    "AlphaOption",
-    "AnalyserOptions",
-    "Animatable",
-    "AnimationEventInit",
-    "AnimationEventMap",
-    "AnimationFrameProvider",
-    "AnimationPlayState",
-    "AnimationPlaybackEventInit",
-    "AnimationReplaceState",
-    "AppendMode",
-    "ArrayBufferConstructor",
-    "ArrayBufferLike",
-    "ArrayBufferTypes",
-    "ArrayBufferView",
-    "ArrayConstructor",
-    "ArrayLike",
-    "AssignedNodesOptions",
-    "AsyncDisposable",
-    "AsyncDisposableStack",
-    "AsyncDisposableStackConstructor",
-    "AsyncGenerator",
-    "AsyncGeneratorFunction",
-    "AsyncGeneratorFunctionConstructor",
-    "AsyncIterable",
-    "AsyncIterableIterator",
-    "AsyncIterator",
-    "Atomics",
-    "AttestationConveyancePreference",
-    "AudioBufferOptions",
-    "AudioBufferSourceOptions",
-    "AudioConfiguration",
-    "AudioContextLatencyCategory",
-    "AudioContextOptions",
-    "AudioContextState",
-    "AudioNodeOptions",
-    "AudioProcessingEventInit",
-    "AudioScheduledSourceNodeEventMap",
-    "AudioTimestamp",
-    "AudioWorkletNodeEventMap",
-    "AudioWorkletNodeOptions",
-    "AuthenticationExtensionsClientInputs",
-    "AuthenticationExtensionsClientOutputs",
-    "AuthenticatorAttachment",
-    "AuthenticatorSelectionCriteria",
-    "AuthenticatorTransport",
-    "AutoFill",
-    "AutoFillAddressKind",
-    "AutoFillBase",
-    "AutoFillContactField",
-    "AutoFillContactKind",
-    "AutoFillCredentialField",
-    "AutoFillField",
-    "AutoFillNormalField",
-    "AutoFillSection",
-    "AutoKeyword",
-    "AutomationRate",
-    "AvcBitstreamFormat",
-    "AvcEncoderConfig",
-    "Awaited",
-    "BaseAudioContextEventMap",
-    "BigInt64ArrayConstructor",
-    "BigIntConstructor",
-    "BigIntToLocaleStringOptions",
-    "BigInteger",
-    "BigUint64ArrayConstructor",
-    "BinaryData",
-    "BinaryType",
-    "BiquadFilterOptions",
-    "BiquadFilterType",
-    "BlobCallback",
-    "BlobEventInit",
-    "BlobPart",
-    "BlobPropertyBag",
-    "Body",
-    "BodyInit",
-    "BooleanConstructor",
-    "BroadcastChannelEventMap",
-    "BufferEncoding",
-    "BufferSource",
-    "COSEAlgorithmIdentifier",
-    "CSSFontFeatureValuesRule",
-    "CSSKeywordish",
-    "CSSMathOperator",
-    "CSSMatrixComponentOptions",
-    "CSSNumberish",
-    "CSSNumericBaseType",
-    "CSSNumericType",
-    "CSSPerspectiveValue",
-    "CSSStyleSheetInit",
-    "CSSUnparsedSegment",
-    "CacheQueryOptions",
-    "CallableFunction",
-    "CanPlayTypeResult",
-    "CanvasCompositing",
-    "CanvasDirection",
-    "CanvasDrawImage",
-    "CanvasDrawPath",
-    "CanvasFillRule",
-    "CanvasFillStrokeStyles",
-    "CanvasFilters",
-    "CanvasFontKerning",
-    "CanvasFontStretch",
-    "CanvasFontVariantCaps",
-    "CanvasImageData",
-    "CanvasImageSmoothing",
-    "CanvasImageSource",
-    "CanvasLineCap",
-    "CanvasLineJoin",
-    "CanvasPath",
-    "CanvasPathDrawingStyles",
-    "CanvasRect",
-    "CanvasRenderingContext2DSettings",
-    "CanvasShadowStyles",
-    "CanvasState",
-    "CanvasText",
-    "CanvasTextAlign",
-    "CanvasTextBaseline",
-    "CanvasTextDrawingStyles",
-    "CanvasTextRendering",
-    "CanvasTransform",
-    "CanvasUserInterface",
-    "Capitalize",
-    "ChannelCountMode",
-    "ChannelInterpretation",
-    "ChannelMergerOptions",
-    "ChannelSplitterOptions",
-    "CheckVisibilityOptions",
-    "ChildNode",
-    "ClassAccessorDecoratorContext",
-    "ClassAccessorDecoratorResult",
-    "ClassAccessorDecoratorTarget",
-    "ClassDecorator",
-    "ClassDecoratorContext",
-    "ClassFieldDecoratorContext",
-    "ClassGetterDecoratorContext",
-    "ClassMemberDecoratorContext",
-    "ClassMethodDecoratorContext",
-    "ClassSetterDecoratorContext",
-    "ClientQueryOptions",
-    "ClientRect",
-    "ClientTypes",
-    "ClipboardEventInit",
-    "ClipboardItemData",
-    "ClipboardItemOptions",
-    "ClipboardItems",
-    "CloseEventInit",
-    "CodecState",
-    "ColorGamut",
-    "ColorSpaceConversion",
-    "CompositeOperation",
-    "CompositeOperationOrAuto",
-    "CompositionEventInit",
-    "CompressionFormat",
-    "ComputedEffectTiming",
-    "ComputedKeyframe",
-    "ConcatArray",
-    "Console",
-    "ConstantSourceOptions",
-    "ConstrainBoolean",
-    "ConstrainBooleanParameters",
-    "ConstrainDOMString",
-    "ConstrainDOMStringParameters",
-    "ConstrainDouble",
-    "ConstrainDoubleRange",
-    "ConstrainULong",
-    "ConstrainULongRange",
-    "ConstructorParameters",
-    "ConvolverOptions",
-    "CredentialCreationOptions",
-    "CredentialMediationRequirement",
-    "CredentialPropertiesOutput",
-    "CredentialRequestOptions",
-    "CryptoKeyPair",
-    "CustomElementConstructor",
-    "CustomEventInit",
-    "DOMHighResTimeStamp",
-    "DOMMatrix2DInit",
-    "DOMMatrixInit",
-    "DOMParserSupportedType",
-    "DOMPointInit",
-    "DOMQuadInit",
-    "DOMRectInit",
-    "DataViewConstructor",
-    "DateConstructor",
-    "DecodeErrorCallback",
-    "DecodeSuccessCallback",
-    "DecoratorContext",
-    "DecoratorMetadata",
-    "DecoratorMetadataObject",
-    "DelayOptions",
-    "DeviceMotionEventAccelerationInit",
-    "DeviceMotionEventInit",
-    "DeviceMotionEventRotationRateInit",
-    "DeviceOrientationEventInit",
-    "DirectionSetting",
-    "DisplayCaptureSurfaceType",
-    "DisplayMediaStreamOptions",
-    "Disposable",
-    "DisposableStack",
-    "DisposableStackConstructor",
-    "DistanceModelType",
-    "DocumentEventMap",
-    "DocumentOrShadowRoot",
-    "DocumentReadyState",
-    "DocumentTimelineOptions",
-    "DocumentVisibilityState",
-    "DoubleRange",
-    "DragEventInit",
-    "DynamicsCompressorOptions",
-    "EXT_blend_minmax",
-    "EXT_color_buffer_float",
-    "EXT_color_buffer_half_float",
-    "EXT_float_blend",
-    "EXT_frag_depth",
-    "EXT_sRGB",
-    "EXT_shader_texture_lod",
-    "EXT_texture_compression_bptc",
-    "EXT_texture_compression_rgtc",
-    "EXT_texture_filter_anisotropic",
-    "EXT_texture_norm16",
-    "EcKeyAlgorithm",
-    "EcKeyGenParams",
-    "EcKeyImportParams",
-    "EcdhKeyDeriveParams",
-    "EcdsaParams",
-    "EffectTiming",
-    "ElementCSSInlineStyle",
-    "ElementContentEditable",
-    "ElementCreationOptions",
-    "ElementDefinitionOptions",
-    "ElementEventMap",
-    "ElementTagNameMap",
-    "EncodedVideoChunkInit",
-    "EncodedVideoChunkMetadata",
-    "EncodedVideoChunkOutputCallback",
-    "EncodedVideoChunkType",
-    "EndOfStreamError",
-    "EndingType",
-    "EpochTimeStamp",
-    "ErrorCallback",
-    "ErrorConstructor",
-    "ErrorEventInit",
-    "ErrorOptions",
-    "EvalErrorConstructor",
-    "EventInit",
-    "EventListener",
-    "EventListenerObject",
-    "EventListenerOptions",
-    "EventListenerOrEventListenerObject",
-    "EventModifierInit",
-    "EventSourceEventMap",
-    "EventSourceInit",
-    "Exclude",
-    "Extract",
-    "FileCallback",
-    "FilePropertyBag",
-    "FileReaderEventMap",
-    "FileSystem",
-    "FileSystemCreateWritableOptions",
-    "FileSystemDirectoryEntry",
-    "FileSystemDirectoryReader",
-    "FileSystemEntriesCallback",
-    "FileSystemEntry",
-    "FileSystemEntryCallback",
-    "FileSystemFileEntry",
-    "FileSystemFlags",
-    "FileSystemGetDirectoryOptions",
-    "FileSystemGetFileOptions",
-    "FileSystemHandleKind",
-    "FileSystemRemoveOptions",
-    "FileSystemWriteChunkType",
-    "FillMode",
-    "FinalizationRegistryConstructor",
-    "FlatArray",
-    "Float32ArrayConstructor",
-    "Float32List",
-    "Float64ArrayConstructor",
-    "FocusEventInit",
-    "FocusOptions",
-    "FontDisplay",
-    "FontFaceDescriptors",
-    "FontFaceLoadStatus",
-    "FontFaceSet",
-    "FontFaceSetEventMap",
-    "FontFaceSetLoadEventInit",
-    "FontFaceSetLoadStatus",
-    "FontFaceSource",
-    "FormDataEntryValue",
-    "FormDataEventInit",
-    "FrameRequestCallback",
-    "FullscreenNavigationUI",
-    "FullscreenOptions",
-    "FunctionConstructor",
-    "FunctionStringCallback",
-    "GLbitfield",
-    "GLboolean",
-    "GLclampf",
-    "GLenum",
-    "GLfloat",
-    "GLint",
-    "GLint64",
-    "GLintptr",
-    "GLsizei",
-    "GLsizeiptr",
-    "GLuint",
-    "GLuint64",
-    "GainOptions",
-    "GamepadEffectParameters",
-    "GamepadEventInit",
-    "GamepadHapticActuatorType",
-    "GamepadHapticEffectType",
-    "GamepadHapticsResult",
-    "GamepadMappingType",
-    "Generator",
-    "GeneratorFunction",
-    "GeneratorFunctionConstructor",
-    "GenericTransformStream",
-    "GetAnimationsOptions",
-    "GetNotificationOptions",
-    "GetRootNodeOptions",
-    "GlobalCompositeOperation",
-    "GlobalEventHandlers",
-    "GlobalEventHandlersEventMap",
-    "HTMLBodyElementEventMap",
-    "HTMLCollectionBase",
-    "HTMLCollectionOf",
-    "HTMLElementDeprecatedTagNameMap",
-    "HTMLElementEventMap",
-    "HTMLElementTagNameMap",
-    "HTMLFrameSetElementEventMap",
-    "HTMLHyperlinkElementUtils",
-    "HTMLMediaElementEventMap",
-    "HTMLOrSVGElement",
-    "HTMLOrSVGImageElement",
-    "HTMLOrSVGScriptElement",
-    "HTMLTableDataCellElement",
-    "HTMLTableHeaderCellElement",
-    "HTMLVideoElementEventMap",
-    "HardwareAcceleration",
-    "HashAlgorithmIdentifier",
-    "HashChangeEventInit",
-    "HdrMetadataType",
-    "HeadersInit",
-    "HkdfParams",
-    "HmacImportParams",
-    "HmacKeyAlgorithm",
-    "HmacKeyGenParams",
-    "IArguments",
-    "IDBCursorDirection",
-    "IDBDatabaseEventMap",
-    "IDBDatabaseInfo",
-    "IDBIndexParameters",
-    "IDBObjectStoreParameters",
-    "IDBOpenDBRequestEventMap",
-    "IDBRequestEventMap",
-    "IDBRequestReadyState",
-    "IDBTransactionDurability",
-    "IDBTransactionEventMap",
-    "IDBTransactionMode",
-    "IDBTransactionOptions",
-    "IDBValidKey",
-    "IDBVersionChangeEventInit",
-    "IIRFilterOptions",
-    "IdleRequestCallback",
-    "IdleRequestOptions",
-    "ImageBitmapOptions",
-    "ImageBitmapRenderingContextSettings",
-    "ImageBitmapSource",
-    "ImageDataSettings",
-    "ImageEncodeOptions",
-    "ImageOrientation",
-    "ImageSmoothingQuality",
-    "ImportAssertions",
-    "ImportCallOptions",
-    "ImportMeta",
-    "InnerHTML",
-    "InputEventInit",
-    "InsertPosition",
-    "InstanceType",
-    "Int16ArrayConstructor",
-    "Int32ArrayConstructor",
-    "Int32List",
-    "Int8ArrayConstructor",
-    "IntersectionObserverCallback",
-    "IntersectionObserverEntryInit",
-    "IntersectionObserverInit",
-    "Iterable",
-    "IterableIterator",
-    "IterationCompositeOperation",
-    "Iterator",
-    "IteratorResult",
-    "IteratorReturnResult",
-    "IteratorYieldResult",
-    "JsonWebKey",
-    "KHR_parallel_shader_compile",
-    "KeyAlgorithm",
-    "KeyFormat",
-    "KeyType",
-    "KeyUsage",
-    "KeyboardEventInit",
-    "Keyframe",
-    "KeyframeAnimationOptions",
-    "KeyframeEffectOptions",
-    "LatencyMode",
-    "LineAlignSetting",
-    "LineAndPositionSetting",
-    "LinkStyle",
-    "LockGrantedCallback",
-    "LockInfo",
-    "LockManagerSnapshot",
-    "LockMode",
-    "LockOptions",
-    "Lowercase",
-    "MIDIAccessEventMap",
-    "MIDIConnectionEventInit",
-    "MIDIInputEventMap",
-    "MIDIMessageEventInit",
-    "MIDIOptions",
-    "MIDIPortConnectionState",
-    "MIDIPortDeviceState",
-    "MIDIPortEventMap",
-    "MIDIPortType",
-    "MapConstructor",
-    "MathMLElementEventMap",
-    "MathMLElementTagNameMap",
-    "MediaCapabilitiesDecodingInfo",
-    "MediaCapabilitiesEncodingInfo",
-    "MediaCapabilitiesInfo",
-    "MediaConfiguration",
-    "MediaDecodingConfiguration",
-    "MediaDecodingType",
-    "MediaDeviceKind",
-    "MediaDevicesEventMap",
-    "MediaElementAudioSourceOptions",
-    "MediaEncodingConfiguration",
-    "MediaEncodingType",
-    "MediaEncryptedEventInit",
-    "MediaImage",
-    "MediaKeyMessageEventInit",
-    "MediaKeyMessageType",
-    "MediaKeySessionClosedReason",
-    "MediaKeySessionEventMap",
-    "MediaKeySessionType",
-    "MediaKeyStatus",
-    "MediaKeySystemConfiguration",
-    "MediaKeySystemMediaCapability",
-    "MediaKeysRequirement",
-    "MediaMetadataInit",
-    "MediaPositionState",
-    "MediaProvider",
-    "MediaQueryListEventInit",
-    "MediaQueryListEventMap",
-    "MediaRecorderEventMap",
-    "MediaRecorderOptions",
-    "MediaSessionAction",
-    "MediaSessionActionDetails",
-    "MediaSessionActionHandler",
-    "MediaSessionPlaybackState",
-    "MediaSourceEventMap",
-    "MediaStreamAudioSourceOptions",
-    "MediaStreamConstraints",
-    "MediaStreamEventMap",
-    "MediaStreamTrackEventInit",
-    "MediaStreamTrackEventMap",
-    "MediaStreamTrackState",
-    "MediaTrackCapabilities",
-    "MediaTrackConstraintSet",
-    "MediaTrackConstraints",
-    "MediaTrackSettings",
-    "MediaTrackSupportedConstraints",
-    "MessageEventInit",
-    "MessageEventSource",
-    "MessagePortEventMap",
-    "MethodDecorator",
-    "MouseEventInit",
-    "MultiCacheQueryOptions",
-    "MutationCallback",
-    "MutationObserverInit",
-    "MutationRecordType",
-    "NamedCurve",
-    "NavigationPreloadState",
-    "NavigationTimingType",
-    "NavigatorAutomationInformation",
-    "NavigatorBadge",
-    "NavigatorConcurrentHardware",
-    "NavigatorContentUtils",
-    "NavigatorCookies",
-    "NavigatorID",
-    "NavigatorLanguage",
-    "NavigatorLocks",
-    "NavigatorOnLine",
-    "NavigatorPlugins",
-    "NavigatorStorage",
-    "NewableFunction",
-    "NoInfer",
-    "NodeJS",
-    "NodeListOf",
-    "NodeRequire",
-    "NonDocumentTypeChildNode",
-    "NonElementParentNode",
-    "NonNullable",
-    "NotificationAction",
-    "NotificationDirection",
-    "NotificationEventMap",
-    "NotificationOptions",
-    "NotificationPermission",
-    "NotificationPermissionCallback",
-    "NumberConstructor",
-    "OES_draw_buffers_indexed",
-    "OES_element_index_uint",
-    "OES_fbo_render_mipmap",
-    "OES_standard_derivatives",
-    "OES_texture_float",
-    "OES_texture_float_linear",
-    "OES_texture_half_float",
-    "OES_texture_half_float_linear",
-    "OES_vertex_array_object",
-    "OVR_multiview2",
-    "ObjectConstructor",
-    "OfflineAudioCompletionEventInit",
-    "OfflineAudioContextEventMap",
-    "OfflineAudioContextOptions",
-    "OffscreenCanvasEventMap",
-    "OffscreenRenderingContext",
-    "OffscreenRenderingContextId",
-    "Omit",
-    "OmitThisParameter",
-    "OnBeforeUnloadEventHandler",
-    "OnBeforeUnloadEventHandlerNonNull",
-    "OnErrorEventHandler",
-    "OnErrorEventHandlerNonNull",
-    "OptionalEffectTiming",
-    "OptionalPostfixToken",
-    "OptionalPrefixToken",
-    "OrientationType",
-    "OscillatorOptions",
-    "OscillatorType",
-    "OverSampleType",
-    "PageTransitionEventInit",
-    "PannerOptions",
-    "PanningModelType",
-    "ParameterDecorator",
-    "Parameters",
-    "ParentNode",
-    "Partial",
-    "PaymentComplete",
-    "PaymentCurrencyAmount",
-    "PaymentDetailsBase",
-    "PaymentDetailsInit",
-    "PaymentDetailsModifier",
-    "PaymentDetailsUpdate",
-    "PaymentItem",
-    "PaymentMethodChangeEventInit",
-    "PaymentMethodData",
-    "PaymentRequestEventMap",
-    "PaymentRequestUpdateEventInit",
-    "PaymentValidationErrors",
-    "Pbkdf2Params",
-    "PerformanceEntryList",
-    "PerformanceEventMap",
-    "PerformanceMarkOptions",
-    "PerformanceMeasureOptions",
-    "PerformanceObserverCallback",
-    "PerformanceObserverInit",
-    "PeriodicWaveConstraints",
-    "PeriodicWaveOptions",
-    "PermissionDescriptor",
-    "PermissionName",
-    "PermissionState",
-    "PermissionStatusEventMap",
-    "Pick",
-    "PictureInPictureEventInit",
-    "PictureInPictureWindowEventMap",
-    "PlaneLayout",
-    "PlaybackDirection",
-    "PointerEventInit",
-    "PopStateEventInit",
-    "PopoverInvokerElement",
-    "PositionAlignSetting",
-    "PositionCallback",
-    "PositionErrorCallback",
-    "PositionOptions",
-    "PredefinedColorSpace",
-    "PremultiplyAlpha",
-    "PresentationStyle",
-    "ProgressEventInit",
-    "PromiseConstructor",
-    "PromiseConstructorLike",
-    "PromiseFulfilledResult",
-    "PromiseLike",
-    "PromiseRejectedResult",
-    "PromiseRejectionEventInit",
-    "PromiseSettledResult",
-    "PropertyDecorator",
-    "PropertyDefinition",
-    "PropertyDescriptor",
-    "PropertyDescriptorMap",
-    "PropertyIndexedKeyframes",
-    "PropertyKey",
-    "ProxyConstructor",
-    "ProxyHandler",
-    "PublicKeyCredentialCreationOptions",
-    "PublicKeyCredentialDescriptor",
-    "PublicKeyCredentialEntity",
-    "PublicKeyCredentialParameters",
-    "PublicKeyCredentialRequestOptions",
-    "PublicKeyCredentialRpEntity",
-    "PublicKeyCredentialType",
-    "PublicKeyCredentialUserEntity",
-    "PushEncryptionKeyName",
-    "PushSubscriptionJSON",
-    "PushSubscriptionOptionsInit",
-    "QueuingStrategy",
-    "QueuingStrategyInit",
-    "QueuingStrategySize",
-    "RTCAnswerOptions",
-    "RTCBundlePolicy",
-    "RTCCertificateExpiration",
-    "RTCConfiguration",
-    "RTCDTMFSenderEventMap",
-    "RTCDTMFToneChangeEventInit",
-    "RTCDataChannelEventInit",
-    "RTCDataChannelEventMap",
-    "RTCDataChannelInit",
-    "RTCDataChannelState",
-    "RTCDegradationPreference",
-    "RTCDtlsFingerprint",
-    "RTCDtlsTransportEventMap",
-    "RTCDtlsTransportState",
-    "RTCEncodedAudioFrameMetadata",
-    "RTCEncodedVideoFrameMetadata",
-    "RTCEncodedVideoFrameType",
-    "RTCErrorDetailType",
-    "RTCErrorEventInit",
-    "RTCErrorInit",
-    "RTCIceCandidateInit",
-    "RTCIceCandidatePair",
-    "RTCIceCandidatePairStats",
-    "RTCIceCandidateType",
-    "RTCIceComponent",
-    "RTCIceConnectionState",
-    "RTCIceGathererState",
-    "RTCIceGatheringState",
-    "RTCIceProtocol",
-    "RTCIceServer",
-    "RTCIceTcpCandidateType",
-    "RTCIceTransportEventMap",
-    "RTCIceTransportPolicy",
-    "RTCIceTransportState",
-    "RTCInboundRtpStreamStats",
-    "RTCLocalSessionDescriptionInit",
-    "RTCOfferAnswerOptions",
-    "RTCOfferOptions",
-    "RTCOutboundRtpStreamStats",
-    "RTCPeerConnectionErrorCallback",
-    "RTCPeerConnectionEventMap",
-    "RTCPeerConnectionIceErrorEventInit",
-    "RTCPeerConnectionIceEventInit",
-    "RTCPeerConnectionState",
-    "RTCPriorityType",
-    "RTCReceivedRtpStreamStats",
-    "RTCRtcpMuxPolicy",
-    "RTCRtcpParameters",
-    "RTCRtpCapabilities",
-    "RTCRtpCodec",
-    "RTCRtpCodecCapability",
-    "RTCRtpCodecParameters",
-    "RTCRtpCodingParameters",
-    "RTCRtpContributingSource",
-    "RTCRtpEncodingParameters",
-    "RTCRtpHeaderExtensionCapability",
-    "RTCRtpHeaderExtensionParameters",
-    "RTCRtpParameters",
-    "RTCRtpReceiveParameters",
-    "RTCRtpSendParameters",
-    "RTCRtpStreamStats",
-    "RTCRtpSynchronizationSource",
-    "RTCRtpTransceiverDirection",
-    "RTCRtpTransceiverInit",
-    "RTCSctpTransportEventMap",
-    "RTCSctpTransportState",
-    "RTCSdpType",
-    "RTCSentRtpStreamStats",
-    "RTCSessionDescriptionCallback",
-    "RTCSessionDescriptionInit",
-    "RTCSignalingState",
-    "RTCStats",
-    "RTCStatsIceCandidatePairState",
-    "RTCStatsType",
-    "RTCTrackEventInit",
-    "RTCTransportStats",
-    "RangeErrorConstructor",
-    "ReadableStreamController",
-    "ReadableStreamGenericReader",
-    "ReadableStreamGetReaderOptions",
-    "ReadableStreamReadDoneResult",
-    "ReadableStreamReadResult",
-    "ReadableStreamReadValueResult",
-    "ReadableStreamReader",
-    "ReadableStreamReaderMode",
-    "ReadableStreamType",
-    "ReadableWritablePair",
-    "Readonly",
-    "ReadonlyArray",
-    "ReadonlyMap",
-    "ReadonlySet",
-    "ReadyState",
-    "Record",
-    "RecordingState",
-    "ReferenceErrorConstructor",
-    "ReferrerPolicy",
-    "RegExpConstructor",
-    "RegExpExecArray",
-    "RegExpIndicesArray",
-    "RegExpMatchArray",
-    "RegistrationOptions",
-    "RemotePlaybackAvailabilityCallback",
-    "RemotePlaybackEventMap",
-    "RemotePlaybackState",
-    "RenderingContext",
-    "Report",
-    "ReportBody",
-    "ReportList",
-    "ReportingObserverCallback",
-    "ReportingObserverOptions",
-    "RequestCache",
-    "RequestCredentials",
-    "RequestDestination",
-    "RequestInfo",
-    "RequestInit",
-    "RequestMode",
-    "RequestRedirect",
-    "Required",
-    "ResidentKeyRequirement",
-    "ResizeObserverBoxOptions",
-    "ResizeObserverCallback",
-    "ResizeObserverOptions",
-    "ResizeQuality",
-    "ResponseInit",
-    "ResponseType",
-    "ReturnType",
-    "RsaHashedImportParams",
-    "RsaHashedKeyAlgorithm",
-    "RsaHashedKeyGenParams",
-    "RsaKeyAlgorithm",
-    "RsaKeyGenParams",
-    "RsaOaepParams",
-    "RsaOtherPrimesInfo",
-    "RsaPssParams",
-    "SVGAnimatedPoints",
-    "SVGBoundingBoxOptions",
-    "SVGElementEventMap",
-    "SVGElementTagNameMap",
-    "SVGFilterPrimitiveStandardAttributes",
-    "SVGFitToViewBox",
-    "SVGSVGElementEventMap",
-    "SVGTests",
-    "SVGURIReference",
-    "ScreenOrientationEventMap",
-    "ScriptProcessorNodeEventMap",
-    "ScrollBehavior",
-    "ScrollIntoViewOptions",
-    "ScrollLogicalPosition",
-    "ScrollOptions",
-    "ScrollRestoration",
-    "ScrollSetting",
-    "ScrollToOptions",
-    "SecurityPolicyViolationEventDisposition",
-    "SecurityPolicyViolationEventInit",
-    "SelectionMode",
-    "ServiceWorkerContainerEventMap",
-    "ServiceWorkerEventMap",
-    "ServiceWorkerRegistrationEventMap",
-    "ServiceWorkerState",
-    "ServiceWorkerUpdateViaCache",
-    "SetConstructor",
-    "ShadowRootEventMap",
-    "ShadowRootInit",
-    "ShadowRootMode",
-    "ShareData",
-    "SharedArrayBuffer",
-    "SharedArrayBufferConstructor",
-    "SlotAssignmentMode",
-    "Slottable",
-    "SourceBufferEventMap",
-    "SourceBufferListEventMap",
-    "SpeechRecognitionAlternative",
-    "SpeechRecognitionResult",
-    "SpeechRecognitionResultList",
-    "SpeechSynthesis",
-    "SpeechSynthesisErrorCode",
-    "SpeechSynthesisErrorEventInit",
-    "SpeechSynthesisEventInit",
-    "SpeechSynthesisEventMap",
-    "SpeechSynthesisUtteranceEventMap",
-    "SpeechSynthesisVoice",
-    "StaticRangeInit",
-    "StereoPannerOptions",
-    "StorageEstimate",
-    "StorageEventInit",
-    "StreamPipeOptions",
-    "StringConstructor",
-    "StructuredSerializeOptions",
-    "StyleMedia",
-    "SubmitEventInit",
-    "SuppressedError",
-    "SuppressedErrorConstructor",
-    "SymbolConstructor",
-    "SyntaxErrorConstructor",
-    "TemplateStringsArray",
-    "TexImageSource",
-    "TextDecodeOptions",
-    "TextDecoderCommon",
-    "TextDecoderOptions",
-    "TextEncoderCommon",
-    "TextEncoderEncodeIntoResult",
-    "TextTrackCueEventMap",
-    "TextTrackEventMap",
-    "TextTrackKind",
-    "TextTrackListEventMap",
-    "TextTrackMode",
-    "Thenable",
-    "ThisParameterType",
-    "ThisType",
-    "TimerHandler",
-    "ToggleEventInit",
-    "TouchEventInit",
-    "TouchInit",
-    "TouchType",
-    "TrackEventInit",
-    "TransferFunction",
-    "Transferable",
-    "Transformer",
-    "TransformerFlushCallback",
-    "TransformerStartCallback",
-    "TransformerTransformCallback",
-    "TransitionEventInit",
-    "TypeErrorConstructor",
-    "TypedPropertyDescriptor",
-    "UIEventInit",
-    "ULongRange",
-    "URIErrorConstructor",
-    "Uint16ArrayConstructor",
-    "Uint32ArrayConstructor",
-    "Uint32List",
-    "Uint8ArrayConstructor",
-    "Uint8ClampedArrayConstructor",
-    "Uncapitalize",
-    "UnderlyingByteSource",
-    "UnderlyingDefaultSource",
-    "UnderlyingSink",
-    "UnderlyingSinkAbortCallback",
-    "UnderlyingSinkCloseCallback",
-    "UnderlyingSinkStartCallback",
-    "UnderlyingSinkWriteCallback",
-    "UnderlyingSource",
-    "UnderlyingSourceCancelCallback",
-    "UnderlyingSourcePullCallback",
-    "UnderlyingSourceStartCallback",
-    "Uppercase",
-    "UserVerificationRequirement",
-    "VTTRegion",
-    "ValidityStateFlags",
-    "VibratePattern",
-    "VideoColorPrimaries",
-    "VideoColorSpaceInit",
-    "VideoConfiguration",
-    "VideoDecoderConfig",
-    "VideoDecoderEventMap",
-    "VideoDecoderInit",
-    "VideoDecoderSupport",
-    "VideoEncoderBitrateMode",
-    "VideoEncoderConfig",
-    "VideoEncoderEncodeOptions",
-    "VideoEncoderEventMap",
-    "VideoEncoderInit",
-    "VideoEncoderSupport",
-    "VideoFacingModeEnum",
-    "VideoFrameBufferInit",
-    "VideoFrameCallbackMetadata",
-    "VideoFrameCopyToOptions",
-    "VideoFrameInit",
-    "VideoFrameOutputCallback",
-    "VideoFrameRequestCallback",
-    "VideoMatrixCoefficients",
-    "VideoPixelFormat",
-    "VideoTransferCharacteristics",
-    "VisualViewportEventMap",
-    "VoidFunction",
-    "WEBGL_color_buffer_float",
-    "WEBGL_compressed_texture_astc",
-    "WEBGL_compressed_texture_etc",
-    "WEBGL_compressed_texture_etc1",
-    "WEBGL_compressed_texture_pvrtc",
-    "WEBGL_compressed_texture_s3tc",
-    "WEBGL_compressed_texture_s3tc_srgb",
-    "WEBGL_debug_renderer_info",
-    "WEBGL_debug_shaders",
-    "WEBGL_depth_texture",
-    "WEBGL_draw_buffers",
-    "WEBGL_lose_context",
-    "WEBGL_multi_draw",
-    "WakeLockSentinelEventMap",
-    "WakeLockType",
-    "WaveShaperOptions",
-    "WeakKey",
-    "WeakKeyTypes",
-    "WeakMapConstructor",
-    "WeakRefConstructor",
-    "WeakSetConstructor",
-    "WebCodecsErrorCallback",
-    "WebGL2RenderingContextBase",
-    "WebGL2RenderingContextOverloads",
-    "WebGLContextAttributes",
-    "WebGLContextEventInit",
-    "WebGLPowerPreference",
-    "WebGLRenderingContextBase",
-    "WebGLRenderingContextOverloads",
-    "WebGLVertexArrayObjectOES",
-    "WebSocketEventMap",
-    "WebTransportCloseInfo",
-    "WebTransportCongestionControl",
-    "WebTransportErrorOptions",
-    "WebTransportErrorSource",
-    "WebTransportHash",
-    "WebTransportOptions",
-    "WebTransportSendStreamOptions",
-    "WheelEventInit",
-    "WindowEventHandlers",
-    "WindowEventHandlersEventMap",
-    "WindowEventMap",
-    "WindowLocalStorage",
-    "WindowOrWorkerGlobalScope",
-    "WindowPostMessageOptions",
-    "WindowProxy",
-    "WindowSessionStorage",
-    "WorkerEventMap",
-    "WorkerOptions",
-    "WorkerType",
-    "WorkletOptions",
-    "WriteCommandType",
-    "WriteParams",
-    "XMLHttpRequestBodyInit",
-    "XMLHttpRequestEventMap",
-    "XMLHttpRequestEventTargetEventMap",
-    "XMLHttpRequestResponseType",
-    "XPathEvaluatorBase",
-    "XPathNSResolver",
-];
-
-#[test]
-fn test_order() {
-    for items in TYPESCRIPT_BUILTIN.windows(2) {
-        assert!(items[0] < items[1], "{} < {}", items[0], items[1]);
-    }
-}
