+++ crates/biome_js_analyze/tests/specs/correctness/noUndeclaredVariables/validWorkerGlobals.ts
@@ -1,0 +1,4 @@
+// web worker
+const syncReader = new FileReaderSync();
+// service worker
+const e = new ExtendableEvent();
